split up astro utils
