+++ src/lib/astro/getAllPostsForSearch.ts
@@ -1,0 +1,36 @@
+import { type CollectionEntry } from 'astro:content'
+import { getAllPosts } from '.'
+
+// helps to reduce DOM size
+
+export async function getAllPostsForSearch() {
+  const allPosts = await getAllPosts()
+
+  const cleaned = await Promise.all(
+    allPosts.map(async (post) => {
+      const imageSrc = (
+        post.data as CollectionEntry<'articles' | 'photos'>['data']
+      ).image
+      // const image = imageSrc
+      //   ? await getImage({
+      //       src: imageSrc,
+      //       width: 300,
+      //       height: 100,
+      //       format: 'png'
+      //     })
+      //   : null
+      return {
+        slug: post.slug,
+        data: {
+          title: post.data.title,
+          tags: post.data.tags,
+          collection: post.collection,
+          lead: post.body.substring(0, 200),
+          image: imageSrc
+        }
+      }
+    })
+  )
+
+  return cleaned
+}
