+++ src/pages/tags/index.astro
@@ -47,3 +47,6 @@ const allTags = await getAllTags(allPosts)
     }
   </ul>
 </LayoutBase>
+ @lib/astro/astro import {getAllPosts} from '@lib/astro/getAllPosts' import {
+  getAllTags
+} from '@lib/astro/AllTags'
